backend/requirements.txt
amqp==2.4.2
astmonkey==0.3.6
attrs==18.2.0
autopep8==1.4
backcall==0.1.0
billiard==3.5.0.5
celery==4.1.1
certifi==2018.8.24
chardet==3.0.4
colorlog==3.1.4
coverage==4.5.3
decorator==4.3.0
Django==2.1.1
django-celery-beat==1.1.1
django-cors-headers==2.4.0
django-environ==0.4.5
django-extensions==2.1.2
django-extra-fields==0.3
django-filter==2.1.0
django-mutpy==0.1.2
django-redis==4.9.0
django-silk==3.0.2
djangorestframework==3.8.2
djangorestframework-filters==0.11.1
djangorestframework-jwt==1.11.0
filetype==1.0.5
flake8==3.5.0
flake8-commas==2.0.0
flake8-mypy==17.8.0
gevent==1.3.6
gprof2dot==2017.9.19
greenlet==0.4.14
gunicorn==19.9.0
idna==2.7
ipython==6.5.0
ipython-genutils==0.2.0
jedi==0.12.1
Jinja2==2.10.1
kombu==4.5.0
lxml==4.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
MutPy==0.6.0
mypy==0.620
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
psycopg2-binary==2.7.5
ptyprocess==0.6.0
pycodestyle==2.3.1
pydot==1.4.1
pyflakes==1.6.0
Pygments==2.2.0
PyJWT==1.7.1
pyparsing==2.4.0
PyPDF2==1.26.0
python-dateutil==2.8.0
pytz==2018.5
PyYAML==5.1
raven==6.9.0
redis==3.2.1
requests==2.19.1
rest-framework-generic-relations==1.2.1
rope==0.11.0
sieve==0.1.9
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.3.0
termcolor==1.1.0
traitlets==4.3.2
typed-ast==1.1.0
urllib3==1.23
vine==1.3.0
wcwidth==0.1.7
Werkzeug==0.14.1
